adjust package names from move
