+++ rc-application/src/main/java/com/transkript/reportcard/business/service/impl/ClassLevelServiceImpl.java
@@ -3,8 +3,8 @@ package com.transkript.reportcard.business.service.impl;
 import com.transkript.reportcard.api.dto.ClassLevelDto;
 import com.transkript.reportcard.api.dto.response.EntityResponse;
 import com.transkript.reportcard.business.mapper.ClassLevelMapper;
-import com.transkript.reportcard.business.service.interf.ClassLevelService;
-import com.transkript.reportcard.business.service.interf.SectionService;
+import com.transkript.reportcard.business.service.i.ClassLevelService;
+import com.transkript.reportcard.business.service.i.SectionService;
 import com.transkript.reportcard.config.constants.EntityName;
 import com.transkript.reportcard.data.entity.ClassLevel;
 import com.transkript.reportcard.data.entity.Section;
