+++ rc-application/src/main/java/com/transkript/reportcard/business/service/impl/ClassListServiceImpl.java
@@ -7,13 +7,13 @@ import com.transkript.reportcard.business.mapper.ClassLevelSubMapper;
 import com.transkript.reportcard.business.mapper.GradeMapper;
 import com.transkript.reportcard.business.mapper.StudentMapper;
 import com.transkript.reportcard.business.mapper.SubjectMapper;
-import com.transkript.reportcard.business.service.interf.AcademicYearService;
-import com.transkript.reportcard.business.service.interf.ClassLevelSubService;
-import com.transkript.reportcard.business.service.interf.ClassListService;
-import com.transkript.reportcard.business.service.interf.SequenceService;
-import com.transkript.reportcard.business.service.interf.StudentApplicationTrialService;
-import com.transkript.reportcard.business.service.interf.SubjectRegistrationService;
-import com.transkript.reportcard.business.service.interf.SubjectService;
+import com.transkript.reportcard.business.service.i.AcademicYearService;
+import com.transkript.reportcard.business.service.i.ClassLevelSubService;
+import com.transkript.reportcard.business.service.i.ClassListService;
+import com.transkript.reportcard.business.service.i.SequenceService;
+import com.transkript.reportcard.business.service.i.StudentApplicationTrialService;
+import com.transkript.reportcard.business.service.i.SubjectRegistrationService;
+import com.transkript.reportcard.business.service.i.SubjectService;
 import com.transkript.reportcard.data.entity.AcademicYear;
 import com.transkript.reportcard.data.entity.ClassLevel;
 import com.transkript.reportcard.data.entity.ClassLevelSub;
