+++ rc-application/src/main/java/com/transkript/reportcard/business/service/impl/GradeServiceImpl.java
@@ -3,10 +3,10 @@ package com.transkript.reportcard.business.service.impl;
 import com.transkript.reportcard.api.dto.GradeDto;
 import com.transkript.reportcard.api.dto.response.EntityResponse;
 import com.transkript.reportcard.business.mapper.GradeMapper;
-import com.transkript.reportcard.business.service.interf.GradeService;
-import com.transkript.reportcard.business.service.interf.SequenceService;
-import com.transkript.reportcard.business.service.interf.StudentApplicationService;
-import com.transkript.reportcard.business.service.interf.SubjectRegistrationService;
+import com.transkript.reportcard.business.service.i.GradeService;
+import com.transkript.reportcard.business.service.i.SequenceService;
+import com.transkript.reportcard.business.service.i.StudentApplicationService;
+import com.transkript.reportcard.business.service.i.SubjectRegistrationService;
 import com.transkript.reportcard.data.entity.Sequence;
 import com.transkript.reportcard.data.entity.SubjectRegistration;
 import com.transkript.reportcard.data.entity.composite.GradeKey;
