+++ rc-application/src/main/java/com/transkript/reportcard/business/service/impl/SchoolServiceImpl.java
@@ -3,7 +3,7 @@ package com.transkript.reportcard.business.service.impl;
 import com.transkript.reportcard.api.dto.SchoolDto;
 import com.transkript.reportcard.api.dto.response.EntityResponse;
 import com.transkript.reportcard.business.mapper.SchoolMapper;
-import com.transkript.reportcard.business.service.interf.SchoolService;
+import com.transkript.reportcard.business.service.i.SchoolService;
 import com.transkript.reportcard.config.constants.EntityName;
 import com.transkript.reportcard.config.constants.ResponseSeverity;
 import com.transkript.reportcard.data.entity.School;
