+++ rc-application/src/main/java/com/transkript/reportcard/business/service/impl/SchoolSettingsServiceImpl.java
@@ -3,10 +3,10 @@ package com.transkript.reportcard.business.service.impl;
 import com.transkript.reportcard.api.dto.SchoolSettingsDto;
 import com.transkript.reportcard.api.dto.response.EntityResponse;
 import com.transkript.reportcard.business.mapper.SchoolSettingsMapper;
-import com.transkript.reportcard.business.service.interf.AcademicYearService;
-import com.transkript.reportcard.business.service.interf.SchoolService;
-import com.transkript.reportcard.business.service.interf.SchoolSettingsService;
-import com.transkript.reportcard.business.service.interf.SequenceService;
+import com.transkript.reportcard.business.service.i.AcademicYearService;
+import com.transkript.reportcard.business.service.i.SchoolService;
+import com.transkript.reportcard.business.service.i.SchoolSettingsService;
+import com.transkript.reportcard.business.service.i.SequenceService;
 import com.transkript.reportcard.config.constants.EntityName;
 import com.transkript.reportcard.config.constants.ResponseSeverity;
 import com.transkript.reportcard.data.entity.AcademicYear;
