+++ rc-application/src/main/java/com/transkript/reportcard/business/service/impl/SequenceServiceImpl.java
@@ -3,8 +3,8 @@ package com.transkript.reportcard.business.service.impl;
 import com.transkript.reportcard.api.dto.SequenceDto;
 import com.transkript.reportcard.api.dto.response.EntityResponse;
 import com.transkript.reportcard.business.mapper.SequenceMapper;
-import com.transkript.reportcard.business.service.interf.SequenceService;
-import com.transkript.reportcard.business.service.interf.TermService;
+import com.transkript.reportcard.business.service.i.SequenceService;
+import com.transkript.reportcard.business.service.i.TermService;
 import com.transkript.reportcard.data.entity.Sequence;
 import com.transkript.reportcard.data.repository.SequenceRepository;
 import com.transkript.reportcard.exception.EntityException;
