+++ rc-application/src/main/java/com/transkript/reportcard/business/service/impl/StudentApplicationServiceImpl.java
@@ -7,10 +7,10 @@ import com.transkript.reportcard.api.dto.response.EntityResponse;
 import com.transkript.reportcard.api.dto.response.StudentApplicationResponse;
 import com.transkript.reportcard.business.mapper.StudentApplicationMapper;
 import com.transkript.reportcard.business.mapper.SubjectRegistrationMapper;
-import com.transkript.reportcard.business.service.interf.AcademicYearService;
-import com.transkript.reportcard.business.service.interf.ClassLevelSubService;
-import com.transkript.reportcard.business.service.interf.StudentApplicationService;
-import com.transkript.reportcard.business.service.interf.StudentService;
+import com.transkript.reportcard.business.service.i.AcademicYearService;
+import com.transkript.reportcard.business.service.i.ClassLevelSubService;
+import com.transkript.reportcard.business.service.i.StudentApplicationService;
+import com.transkript.reportcard.business.service.i.StudentService;
 import com.transkript.reportcard.config.constants.EntityName;
 import com.transkript.reportcard.data.entity.AcademicYear;
 import com.transkript.reportcard.data.entity.ClassLevelSub;
