+++ rc-application/src/main/java/com/transkript/reportcard/business/service/impl/StudentServiceImpl.java
@@ -3,8 +3,8 @@ package com.transkript.reportcard.business.service.impl;
 import com.transkript.reportcard.api.dto.StudentDto;
 import com.transkript.reportcard.api.dto.response.EntityResponse;
 import com.transkript.reportcard.business.mapper.StudentMapper;
-import com.transkript.reportcard.business.service.interf.SchoolSettingsService;
-import com.transkript.reportcard.business.service.interf.StudentService;
+import com.transkript.reportcard.business.service.i.SchoolSettingsService;
+import com.transkript.reportcard.business.service.i.StudentService;
 import com.transkript.reportcard.business.util.SchoolUtil;
 import com.transkript.reportcard.business.util.SettingsUtil;
 import com.transkript.reportcard.data.entity.Student;
