+++ rc-application/src/main/java/com/transkript/reportcard/business/service/impl/SubjectRegistrationServiceImpl.java
@@ -3,10 +3,10 @@ package com.transkript.reportcard.business.service.impl;
 import com.transkript.reportcard.api.dto.SubjectRegistrationDto;
 import com.transkript.reportcard.api.dto.response.EntityResponse;
 import com.transkript.reportcard.business.mapper.SubjectRegistrationMapper;
-import com.transkript.reportcard.business.service.interf.StudentApplicationService;
-import com.transkript.reportcard.business.service.interf.StudentApplicationTrialService;
-import com.transkript.reportcard.business.service.interf.SubjectRegistrationService;
-import com.transkript.reportcard.business.service.interf.SubjectService;
+import com.transkript.reportcard.business.service.i.StudentApplicationService;
+import com.transkript.reportcard.business.service.i.StudentApplicationTrialService;
+import com.transkript.reportcard.business.service.i.SubjectRegistrationService;
+import com.transkript.reportcard.business.service.i.SubjectService;
 import com.transkript.reportcard.config.constants.EntityName;
 import com.transkript.reportcard.data.entity.Subject;
 import com.transkript.reportcard.data.entity.SubjectRegistration;
