+++ rc-application/src/main/java/com/transkript/reportcard/business/service/impl/SubjectServiceImpl.java
@@ -3,8 +3,8 @@ package com.transkript.reportcard.business.service.impl;
 import com.transkript.reportcard.api.dto.SubjectDto;
 import com.transkript.reportcard.api.dto.response.EntityResponse;
 import com.transkript.reportcard.business.mapper.SubjectMapper;
-import com.transkript.reportcard.business.service.interf.SectionService;
-import com.transkript.reportcard.business.service.interf.SubjectService;
+import com.transkript.reportcard.business.service.i.SectionService;
+import com.transkript.reportcard.business.service.i.SubjectService;
 import com.transkript.reportcard.data.entity.Section;
 import com.transkript.reportcard.data.entity.Subject;
 import com.transkript.reportcard.data.repository.SubjectRepository;
