+++ rc-application/src/main/java/com/transkript/reportcard/business/service/impl/TermServiceImpl.java
@@ -3,8 +3,8 @@ package com.transkript.reportcard.business.service.impl;
 import com.transkript.reportcard.api.dto.TermDto;
 import com.transkript.reportcard.api.dto.response.EntityResponse;
 import com.transkript.reportcard.business.mapper.TermMapper;
-import com.transkript.reportcard.business.service.interf.SchoolService;
-import com.transkript.reportcard.business.service.interf.TermService;
+import com.transkript.reportcard.business.service.i.SchoolService;
+import com.transkript.reportcard.business.service.i.TermService;
 import com.transkript.reportcard.data.entity.Term;
 import com.transkript.reportcard.data.repository.TermRepository;
 import com.transkript.reportcard.exception.EntityException;
