+++ rc-application/src/main/java/com/transkript/reportcard/business/service/impl/UserServiceImpl.java
@@ -3,8 +3,7 @@ package com.transkript.reportcard.business.service.impl;
 import com.transkript.reportcard.api.dto.UserDto;
 import com.transkript.reportcard.api.dto.response.EntityResponse;
 import com.transkript.reportcard.business.mapper.UserMapper;
-import com.transkript.reportcard.business.service.interf.UserService;
-import com.transkript.reportcard.business.util.JwtUtil;
+import com.transkript.reportcard.business.service.i.UserService;
 import com.transkript.reportcard.data.entity.User;
 import com.transkript.reportcard.data.repository.UserRepository;
 import com.transkript.reportcard.exception.EntityException;
